refactor: remove unused actorDto imports across multiple files
