+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/mapper/search/ActorSearchResponseMapper.kt
@@ -3,8 +3,8 @@ package com.baghdad.remoteDataSource.mapper.search
 import com.baghdad.remoteDataSource.response.search.ActorSearchResponse
 import com.baghdad.remoteDataSource.util.getNextKey
 import com.baghdad.remoteDataSource.util.getPreviousKey
+import com.baghdad.repository.model.ActorDto
 import com.baghdad.repository.model.PagedResultDto
-import com.baghdad.repository.model.actor.ActorDto
 
 fun ActorSearchResponse.toPagedActorDtos() = PagedResultDto(
     data = results?.mapNotNull { it?.toActorDto() } ?: emptyList(),
